Isolate pull selector tests and asign to RequireCloudant test category
